// src/shared/dataprovider/api/model/userDTO.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StatusEnum } from './statusEnum';

export interface UserDTO { 
    Id?: number;
    Username?: string;
    Name?: string;
    Email?: string;
    Status?: StatusEnum;
    Token?: string;
    LastLoginIP?: string;
    LastLoginDate?: Date;
    CreatedDate?: Date;
    LastUpdatedDate?: Date;
    PasswordDidChanged?: boolean;
    PasswordLastUpdated?: Date;
    ClientId?: number;
    RoleId?:number;
    ClientName?:string;
}